Refactor to make the tool easier to develop
Tested on all AM2xxx XMLs, no change in output
